Struttura Dati per Basic Event Firing Time: - Si può creare automaticamente uno scenario a partire da un sistema specificato a livello Meta mediante il metodo InitializeScenarioFromSystem. In questo modo le occorrenze dei fault vengono campionate a partire dalla loro pdf. -Si possono specificare delle occorrenze deterministiche di Fault, Error e Failure tramite il metodo AddEvent, esplicitando il timestamp in cui avvengono. In questo modo, l'occorrenza di un Fault è di base campionata dalla PDF, altrimenti è rimpiazzata con l'istante specificato dall'utente. -Bug fix

+++ src/main/java/it/unifi/stlab/fault2failure/knowledge/utils/PDFParser.java
@@ -6,7 +6,6 @@ import org.oristool.math.function.Erlang;
 import org.oristool.math.function.GEN;
 import org.oristool.models.stpn.trees.StochasticTransitionFeature;
 
-import java.math.BigDecimal;
 
 public class PDFParser {
     public static StochasticTransitionFeature parseStringToStochasticTransitionFeature(String arisingPDF) {
@@ -102,7 +101,7 @@ public class PDFParser {
                 return StochasticTransitionFeature.newUniformInstance(a,b);
             }
         } else if (realDistribution.getClass().equals(ExponentialDistribution.class)) {
-            return StochasticTransitionFeature.newExponentialInstance(String.valueOf((int)(1/(((ExponentialDistribution) realDistribution).getMean()))));
+            return StochasticTransitionFeature.newExponentialInstance(String.valueOf((1/(((ExponentialDistribution) realDistribution).getMean()))));
         } else if (realDistribution.getClass().equals(GammaDistribution.class)) {
             return StochasticTransitionFeature.newErlangInstance(((int) ((GammaDistribution) realDistribution).getShape()), String.valueOf((int)((GammaDistribution) realDistribution).getScale()));
         } else
